+++ procedure.py
@@ -10,7 +10,7 @@ import os
 def prepare_batch(batch, device):  
     record, label = batch 
     assert label.dim() == 1 and record.dim() == 2, f'{record.shape}, {label.shape}'
-    return (record.to(device), label.unsqueeze(dim=-1).to(device)) 
+    return (record.to(device), label.float().unsqueeze(dim=-1).to(device)) 
 
 def train(model, dataloader, optimizer, criterion, device): 
     model.train()
